% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cuminc_risktable.R
\name{add_cuminc_risktable}
\alias{add_cuminc_risktable}
\title{Add risk table to \code{cuminc()} plot}
\usage{
add_cuminc_risktable(
  cuminc,
  survfit,
  timepts,
  lg,
  numgrps,
  line = 4,
  at = -1,
  col.list = 1
)
}
\arguments{
\item{cuminc}{\code{cmprsk::cuminc()} object}

\item{survfit}{\code{survival::survfit()} object}

\item{timepts}{a numeric vector of time points of the estimates to display.
E.g. c(0,1,2,3,4,5) or seq(0,12,by=2)}

\item{lg}{legend label for each cumulative incidence curve to be displayed.
E.g. c("Male", "Female")}

\item{numgrps}{the number of groups of the stratification variable: 1 is
no stratification, 2 is stratified by a binary variable}

\item{line}{to adjust position of risk table. A lower value will shift
table up, a larger value will shift table down; default is 4}

\item{at}{to adjust position of left margin. Default is -1.}

\item{col.list}{list of colors for legend text Should match the colors
of plot legend. Default is 1 (black).}
}
\description{
Plot cumulative incidence estimates with a risk table and estimates below
the figure.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(cmprsk)
library(survival)

data(pbc)
# recode time and status
pbc$time.y <- pbc$time / 365.25
pbc$status2 <- ifelse(pbc$status > 0, ifelse(pbc$status == 1, 2, 1), 0)

# Example 1 -------------------------------------
# CIR and KM for no strata
cif1 <- cuminc(ftime = pbc$time.y, fstatus = pbc$status2)
km1 <- survfit(Surv(pbc$time.y, pbc$status == 1) ~ 1)

# Plot and add risk table for no strata (numgrps=1)
windows(5, 5)
par(mfrow = c(1, 1),
    mar = c(12.5, 5.7, 2, 2),
    mgp = c(2, 0.65, 0))
plot(cif1,
     curvlab = c("recurred", "died"),
     xlim = c(0, 12), xaxt = "n")
axis(1, at = seq(0, 12, 3))

add_cuminc_risktable(cif1, km1,
                     timepts = seq(0, 12, 3),
                     lg = "",
                     numgrps = 1)

# Example 2 -------------------------------------
cif2 <- cuminc(ftime = pbc$time.y,
               fstatus = pbc$status2,
               group = pbc$sex)
km2 <- survfit(Surv(pbc$time.y, pbc$status == 1) ~ pbc$sex)

# Plot and add risk table for 2 groups (numgrps=2)
windows(5, 5)
par(mfrow = c(1, 1),
    mar = c(12.5, 5.7, 2, 2),
    mgp = c(2, 0.65, 0))

plot(cif2,
     curvlab = c("male", "female", "", ""),
     lty = c(1, 2, 0, 0),
     xlim = c(0, 12),
     xaxt = "n", col = c(1, 3, 0, 0))
axis(1, at = seq(0, 12, 3))

add_cuminc_risktable(cif2, km2,
                     timepts = seq(0, 12, 3),
                     lg = c("male", "female"),
                     numgrps = 2)
\dontshow{\}) # examplesIf}
}
\author{
Meier Hsu
}
