% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_forest_plot.R
\name{as_forest_plot}
\alias{as_forest_plot}
\title{Create Forest Plot}
\usage{
as_forest_plot(
  x,
  col_names = c("estimate", "ci", "p.value"),
  graph.pos = 2,
  boxsize = 0.3,
  title_line_color = "darkblue",
  ...
)
}
\arguments{
\item{x}{a gtsummary object of class \code{"tbl_regression"} or \code{"tbl_uvregression"}}

\item{col_names}{names of columns in \code{x$table_body} to print on the RHS of the
forest plot. Default is \code{c("estimate", "ci", "p.value")}}

\item{graph.pos}{The position of the graph element within the table of text. The
position can be \code{1-(ncol(labeltext) + 1)}. You can also choose set the position
to \code{"left"} or \code{"right"}.}

\item{boxsize}{Override the default box size based on precision}

\item{title_line_color}{color of line that appears above forest plot.
Default is \code{"darkblue"}}

\item{...}{arguments passed to \code{forestplot::forestplot()}}
}
\description{
\lifecycle{experimental}
This function takes a gtsummary table and converts it to a
forest plot using \code{forestplot::forestplot()}.
}
\examples{
library(gtsummary)
library(survival)

# Example 1 ----------------------------------
tbl_uvregression(
  trial[c("response", "age", "grade")],
  method = glm,
  y = response,
  method.args = list(family = binomial),
  exponentiate = TRUE
) \%>\%
as_forest_plot()

# Example 2 ------------------------------------
tbl <-
 coxph(Surv(ttdeath, death) ~ age + marker, trial) \%>\%
 tbl_regression(exponentiate = TRUE) \%>\%
 add_n()

as_forest_plot(tbl, col_names = c("stat_n", "estimate", "ci", "p.value"))

# Example 3 ----------------------------------
tbl \%>\%
modify_cols_merge(
  pattern = "{estimate} ({ci})",
  rows = !is.na(estimate)
) \%>\%
modify_header(estimate = "HR (95\% CI)") \%>\%
as_forest_plot(
  col_names = c("estimate", "p.value"),
  boxsize = 0.2,
  col = forestplot::fpColors(box = "darkred")
)
}
\author{
Christine Zhou
}
