% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ggplot.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Convert gt/gtsummary table to ggplot}
\usage{
as_ggplot(x, ...)
}
\arguments{
\item{x}{gt or gtsummary table}

\item{...}{arguments passed to \code{gt::gtsave()}}
}
\value{
a ggplot object
}
\description{
useful when you want to place a ggplot and gt table side-by-side
}
\examples{
library(gtsummary)
library(ggplot2)
library(patchwork)

# convert gtsummary table to ggplot
tbl <-
  trial \%>\%
  select(age, marker, trt) \%>\%
  tbl_summary(by = trt, missing = "no") \%>\%
  as_ggplot()

# create basic ggplot
gg <-
  trial \%>\%
  ggplot(aes(x = age, y = marker, color = trt)) +
  geom_point()

# stack tables using {patchwork}
gg / tbl
}
