% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_italicise_group_labels.R
\name{bold_italicise_group_labels}
\alias{bold_italicise_group_labels}
\title{Set bold and/or italic style for groups labels in stacked tables}
\usage{
bold_italicise_group_labels(
  x,
  bold = FALSE,
  italics = FALSE,
  print_engine = c("gt", "flextable", "huxtable")
)
}
\arguments{
\item{x}{a gtsummary stacked table}
}
\description{
Set bold and/or italic style for groups labels in stacked tables
}
