% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstfun-package.R
\docType{package}
\name{bstfun-package}
\alias{bstfun}
\alias{bstfun-package}
\title{bstfun: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/bstfun}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
