% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mrn.R
\name{clean_mrn}
\alias{clean_mrn}
\title{Check and Format MRNs}
\usage{
clean_mrn(x, allow_na = FALSE, check_unique = FALSE)
}
\arguments{
\item{x}{vector to be converted and checked to MRN}

\item{allow_na}{logical indicating whether \code{NA} values are accepted.
Default is \code{FALSE}}

\item{check_unique}{Check if MRNs are unique}
}
\value{
character MRN vector
}
\description{
An MRN follows specific rules
\enumerate{
\item Must be character
\item Must contain only numeric components
\item Must be eight characters long and include leading zeros.
}

This function converts numeric MRNs to character and ensures it follows
MRN conventions. Character MRNs can also be passed, and leading zeros
will be appended and checked for consistency.
}
\examples{
1000:1001 \%>\%
  clean_mrn()
}
