% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_table.R
\name{compact_table}
\alias{compact_table}
\title{Compact Table Styling}
\usage{
compact_table(data)
}
\arguments{
\item{data}{a {gt}, {flextable}, {huxtable}, or \code{knitr::kable()} table.}
}
\description{
Apply the same compact styling available from
\code{gtsummary::theme_gtsummary_compact()} to any
{gt}, {flextable}, {huxtable}, or \code{knitr::kable()} table.
\code{knitr::kable()} stlying uses the {kableExtra} package
}
\examples{
compact_table_ex1 <-
  head(trial) \%>\%
  gt::gt() \%>\%
  compact_table()
}
