% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_map.R
\name{count_map}
\alias{count_map}
\title{Check variable derivations}
\usage{
count_map(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{sets of variables to check. variables that are checked together
are included in the same vector. See example below.}
}
\description{
Function assists in checking the values of new, derived variables against
the raw, source variables.
}
\examples{
count_map(
  mtcars,
  c(cyl, am), c(gear, carb)
)
}
