% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Assess pattern of missing data}
\usage{
count_na(data, include = NULL, exclude = NULL)
}
\arguments{
\item{data}{data frame}

\item{include}{character vector of names to include}

\item{exclude}{character vector of names to exclude}
}
\value{
original data frame invisibly returned
}
\description{
Pass a data frame and the missing pattern of all columns in the data frame.
The data frame is returned unmodified.
}
\examples{
trial \%>\% count_na()
}
