% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{reinstall_prior_pkgs}
\alias{tbl_2way_summary}
\alias{gtsummary_butcher}
\alias{gts_add_p_footnotes}
\title{Deprecated functions}
\usage{
reinstall_prior_pkgs(...)

tbl_2way_summary(
  data,
  row,
  col,
  con,
  label = NULL,
  statistic = "{median} ({p25}, {p75})"
)

gtsummary_butcher(x)

gts_add_p_footnotes(x, printer = NULL, index_start = NULL)
}
\description{
\lifecycle{deprecated}
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
