% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{egfr}
\alias{egfr}
\alias{egfr_ckdepi}
\alias{egfr_mdrd}
\title{Calculate eGFR}
\usage{
egfr_ckdepi(creatinine, age, female, aa, label = "eGFR, mL/min/1.73m²")

egfr_mdrd(creatinine, age, female, aa, label = "eGFR, mL/min/1.73m²")
}
\arguments{
\item{creatinine}{serum creatinine level in mg/dL}

\item{age}{patient age}

\item{female}{logical indicating whether patient is female}

\item{aa}{logical indicating whether patient is African-American}

\item{label}{label that will be applied to result,
e.g. \code{attr("label", 'eGFR, mL/min/1.73m\\U00B2')}}
}
\value{
numeric vector
}
\description{
Calculate eGFR
}
\examples{
egfr_mdrd(creatinine = 1.2, age = 60, female = TRUE, aa = TRUE)
egfr_ckdepi(creatinine = 1.2, age = 60, female = TRUE, aa = TRUE)
}
