% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Calculates the mode(s) of a set of values}
\usage{
get_mode(x, moden = 1, quiet = FALSE)
}
\arguments{
\item{x}{A variable or vector (numeric, character or factor)}

\item{moden}{If there are multiple modes, which mode to use. The default is the first mode.}

\item{quiet}{By default, messages are printed if multiple modes are selected. To hide these messages, set \code{quiet} to \code{TRUE}}
}
\value{
A vector of length 1 containing the mode
}
\description{
This function calculates the most common value(s) of a given set
}
\examples{
get_mode(trial$stage)
get_mode(trial$trt)
get_mode(trial$response)
get_mode(trial$grade)
}
