% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalibration.R
\name{ggcalibration}
\alias{ggcalibration}
\title{Model Calibration Plot}
\usage{
ggcalibration(
  data,
  y,
  x,
  n.groups = 10,
  conf.level = 0.95,
  methods = c("exact", "ac", "asymptotic", "wilson", "prop.test", "bayes", "logit",
    "cloglog", "probit"),
  geom_errorbar.args = list(width = 0),
  geom_point.args = list()
)
}
\arguments{
\item{data}{a data frame}

\item{y}{variable name of the outcome coded as 0/1}

\item{x}{variable name of the risk predictions}

\item{n.groups}{number of groups}

\item{conf.level}{level of confidence to be used in the confidence interval}

\item{methods}{method to use to construct the interval.
See \code{\link[binom:binom.confint]{binom::binom.confint()}} for details}

\item{geom_errorbar.args}{named list of arguments that will be passed
to \code{ggplot2::geom_errorbar()}. Default is \code{list(width = 0)}}

\item{geom_point.args}{named list of arguments that will be passed
to \code{ggplot2::geom_point()}. Default is \code{list()}}
}
\value{
ggplot
}
\description{
Assess a model's calibration via a calibration plot.
}
\examples{
glm(response ~ age + marker + grade, trial, family = binomial) \%>\%
  broom::augment(type.predict = "response") \%>\%
  ggcalibration(y = response, x = .fitted, n.groups = 6)
}
