% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_data.R
\name{here_data}
\alias{here_data}
\title{Find your data folder}
\usage{
here_data(..., data_folder_name = "secure_data")
}
\arguments{
\item{...}{\code{[character]}\cr
Path components below the project root, can be empty.}

\item{data_folder_name}{name of data folder. Default is \code{"secure_data"}}
}
\value{
path to data folder
}
\description{
Similar to \code{here::here()} which returns your project directory, \code{here_data()}
returns the path to your current data folder based on the current \code{data_date.txt}
}
\details{
The function expects the user to version their data using a text file indicating
the date the data was last received, and the data to be stored in a corresponding
folder name, e.g. \verb{~\\Project Folder\\secure_data\\2020-01-01}.
}
\examples{
\dontrun{
here_data()
#> "C:\Users\SjobergD\GitHub\My Project\secure_data\2020-01-01"

here_data("Raw Data.xlsx")
#> "C:\Users\SjobergD\GitHub\My Project\secure_data\2020-01-01\Raw Data.xlsx"
}
}
