% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpcc.R
\name{hpcc}
\alias{hpcc}
\alias{hpcc_get_arg}
\alias{hpcc_get_seq_number}
\title{HPCC Functions}
\usage{
hpcc_get_arg()

hpcc_get_seq_number()
}
\description{
Functions for working with the high performance computing cluster
\itemize{
\item \code{hpcc_get_arg()} retrieve character argument passed from terminal,
e.g. \verb{qsubR bootstrap_analysis.R "mets"}
\item \code{hpcc_get_seq_number()} retrieve sequence integer when a sequence of jobs
were submitted from the terminal, e.g \verb{qsubR -a 1-50\%10 bootstrap_analysis.R}
}
}
