% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cuminc.R
\name{plot_cuminc}
\alias{plot_cuminc}
\title{Plot \code{cuminc()} Estimates}
\usage{
plot_cuminc(
  cuminc,
  survfit,
  timepts,
  lg,
  numgrps,
  lgd.at = "topleft",
  line = 4,
  at = -1,
  col.list = 1
)
}
\arguments{
\item{cuminc}{\code{cmprsk::cuminc()} object}

\item{survfit}{\code{survival::survfit()} object}

\item{col.list}{}
}
\description{
Plot \code{cuminc()} Estimates
}
\author{
Meier Hsu
}
