% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_when.R
\name{purrr_when}
\alias{purrr_when}
\title{Slightly modified version of purrr::when. Match/validate a set of conditions
for an object and continue with the action associated with the first valid match.}
\usage{
purrr_when(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A sequence of two-sided
formulas. The left hand side (LHS) determines which values match this
case. The right hand side (RHS) provides the replacement value.
The LHS must evaluate to a logical vector. The RHS does not need to be logical,
and all RHSs are not required to evaluate to the same type of vector.}
}
\value{
The result of the action associated with the first valid match/condition found
}
\description{
purrr_when() serves as a replacement for the when() function, which was
deprecated in purrr 1.0.0.
}
\details{
Copied from purrr::when description:
purrr_when is a flavour of pattern matching (or an if-else abstraction) in which a value
is matched against a sequence of condition-action sets. When a valid
match/condition is found the action is executed and the result of the
action is returned.
}
\examples{
y <- 2
purrr_when(
  y < 8 ~ y + 2,
  y > 8 ~ "y > 8"
)
y <- 10
purrr_when(
  y < 8 ~ y + 2,
  y > 8 ~ "y > 8"
)
y <- 8
purrr_when(
  y < 8 ~ y + 2,
  y > 8 ~ "y > 8"
)
}
\author{
Daniel D Sjoberg
}
