% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_derived_variables.R
\name{set_derived_variables}
\alias{set_derived_variables}
\title{Apply variable labels to data frame}
\usage{
set_derived_variables(data, path, sheet = NULL, drop = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{path}{Path to Derived Variables xls/xlsx file}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{drop}{Logical indicating whether to drop unlabelled variables}
}
\description{
Takes labels from the Derived Variables excel file and applies them to the
passed data frame.
The excel sheet must have columns \code{"varname"} and \code{"label"}.
}
\examples{
\dontrun{
\donttest{
trial \%>\%
  set_derived_variables("derived_variables_sjoberg.xlsx")
}
}
}
\author{
Daniel D. Sjoberg
}
