% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ancova.R
\name{tbl_ancova}
\alias{tbl_ancova}
\title{Table difference between two groups}
\usage{
tbl_ancova(
  data,
  y,
  x,
  formula = "{y} ~ {x}",
  label = NULL,
  method.args = NULL,
  conf.level = 0.95,
  estimate_fun = NULL,
  pvalue_fun = NULL,
  method = stats::lm,
  digits = NULL
)
}
\arguments{
\item{data}{frame to be used in ANCOVA models}

\item{y}{vector of continuous outcome variables. One-way ANOVA/ANCOVA models
will be computed for each outcome.}

\item{x}{string indicating the binary comparison variable}

\item{formula}{String of the model formula.
Uses \link[glue:glue]{glue::glue} syntax. Default is \code{"{y} ~ {x}"}, where \code{{y}}
is the dependent variable, and \code{{x}} represents a single covariate. For a
random intercept model, the formula may be \code{formula = "{y} ~ {x} + (1 | gear)"}.}

\item{label}{List of formulas specifying variables labels,
e.g. \code{list(age ~ "Age, yrs", stage ~ "Path T Stage")}}

\item{method.args}{List of additional arguments passed on to the regression
function defined by \code{method}.}

\item{conf.level}{Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{estimate_fun}{Function to round and format coefficient estimates.
Default is \link[gtsummary]{style_sigfig} when the coefficients are not transformed, and
\link[gtsummary]{style_ratio} when the coefficients have been exponentiated.}

\item{pvalue_fun}{Function to round and format p-values.
Default is \link[gtsummary]{style_pvalue}.
The function must have a numeric vector input (the numeric, exact p-value),
and return a string that is the rounded/formatted p-value (e.g.
\code{pvalue_fun = function(x) style_pvalue(x, digits = 2)} or equivalently,
\code{purrr::partial(style_pvalue, digits = 2)}).}

\item{method}{Regression method (e.g. \link{lm}, \link{glm}, \link[survival:coxph]{survival::coxph}, and more).}

\item{digits}{List of formulas specifying the number of decimal
places to round continuous summary statistics. If not specified,
\code{tbl_summary} guesses an appropriate number of decimals to round statistics.
When multiple statistics are displayed for a single variable, supply a vector
rather than an integer.  For example, if the
statistic being calculated is \code{"{mean} ({sd})"} and you want the mean rounded
to 1 decimal place, and the SD to 2 use \code{digits = list(age ~ c(1, 2))}.}
}
\description{
Returns an ANCOVA table showing the means of two groups and the
difference in means
}
\examples{
gtsummary::trial \%>\%
  tbl_ancova(y = c("age", "marker"), x = "trt")
}
\author{
Daniel D. Sjoberg
}
