% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_split.R
\name{tbl_split}
\alias{tbl_split}
\alias{print.tbl_split}
\title{Split gtsummary Table}
\usage{
tbl_split(x, variables)

\method{print}{tbl_split}(x, ...)
}
\arguments{
\item{x}{gtsummary table}

\item{variables}{variables to split gtsummary table at}

\item{...}{not used}
}
\value{
tbl_split object
}
\description{
Split gtsummary Table
}
\examples{
library(gtsummary)
tbl <-
  tbl_summary(trial) \%>\%
  tbl_split(variables = c(marker, grade))
}
