% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtsummary_bst.R
\name{theme_gtsummary_bst}
\alias{theme_gtsummary_bst}
\title{Set custom gtsummary themes}
\usage{
theme_gtsummary_bst(name = c("hot", "karissa"))
}
\arguments{
\item{name}{string indicating the custom theme to set.}
}
\description{
The function contains the personal themes for members of the department.
Reach out if you're interested in adding your own personal theme!
}
\details{
Visit the \href{http://www.danieldsjoberg.com/gtsummary/articles/themes.html#writing-themes-1}{gtsummary themes vignette} for a full list of preferences that can be set.
}
\examples{
theme_gtsummary_bst("karissa")
}
