% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtsummary_msk.R
\name{theme_gtsummary_msk}
\alias{theme_gtsummary_msk}
\title{Set custom gtsummary themes}
\usage{
theme_gtsummary_msk(name = c("hot"))
}
\arguments{
\item{name}{string indicating the custom theme to set.}
}
\description{
This is a place for any member of the MSK community to add a personal gtsummary theme.
Reach out if you're interested in adding yours!
}
\details{
Visit the \href{http://www.danieldsjoberg.com/gtsummary/articles/themes.html#writing-themes-1}{gtsummary themes vignette} for a full list of preferences that can be set.
}
\examples{
theme_gtsummary_msk("hot")
}
