% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_biostat_file.R
\name{use_biostat_file}
\alias{use_biostat_file}
\alias{use_biostat_gitignore}
\alias{use_biostat_readme}
\title{Write a template file}
\usage{
use_biostat_file(
  name = NULL,
  filename = NULL,
  template = NULL,
  open = interactive()
)

use_biostat_gitignore(filename = NULL, template = NULL)

use_biostat_readme(filename = NULL, template = NULL)
}
\arguments{
\item{name}{Name of file to write.  Not sure of the files available to you?
Run the function without specifying a name, and all files available within the
template will print.}

\item{filename}{Optional argument to specify the name of the file to be written.
Paths/filename is relative to project base (e.g. \code{here::here()})}

\item{template}{Project template.}

\item{open}{If \code{TRUE}, opens the new file.}
}
\description{
Rather than using \code{create_biostat_project()} to start a new project folder, you
may use \code{use_biostat_file()} to write a single file from any project template.
The functions \code{use_biostat_gitignore()} and \code{use_biostat_readme()} are shortcuts for
\code{use_biostat_file("gitignore")} and \code{use_biostat_file("readme")}.
}
\examples{
\donttest{\dontrun{
# create gitignore file
use_project_file("gitignore")
use_project_gitignore()

# create README.md file
use_project_file("readme")
use_project_readme()
}}
}
\seealso{
\code{\link[=create_biostat_project]{create_biostat_project()}}
}
