% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_varnames_as_labels.R
\name{use_varnames_as_labels}
\alias{use_varnames_as_labels}
\title{Use Variable Names as Labels}
\usage{
use_varnames_as_labels(data, caps = NULL, exclude = NULL)
}
\arguments{
\item{data}{a data frame}

\item{caps}{variables to be entirely capitalized}

\item{exclude}{variables to exclude from labeling}
}
\value{
a labeled data frame
}
\description{
Use Variable Names as Labels
}
\examples{
library(gtsummary)

mtcars \%>\%
  select(mpg, cyl, vs, am, carb) \%>\%
  use_varnames_as_labels(caps = c(mpg, vs, am), exclude = cyl) \%>\%
  tbl_summary() \%>\%
  as_kable(format = "simple")
}
