% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_likert.R
\name{add_continuous_stat}
\alias{add_continuous_stat}
\title{Add continuous statistic to gtsummary table}
\usage{
add_continuous_stat(x, ...)
}
\arguments{
\item{x}{Object created from a gtsummary function}

\item{...}{Additional arguments passed to other methods}
}
\description{
Add continuous statistic to gtsummary table
}
\seealso{
\code{\link[=add_continuous_stat.tbl_likert]{add_continuous_stat.tbl_likert()}}
}
