% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_likert.R
\name{add_n.tbl_likert}
\alias{add_n.tbl_likert}
\title{Add column with N}
\usage{
\method{add_n}{tbl_likert}(
  x,
  statistic = "{n}",
  col_label = "**N**",
  footnote = FALSE,
  last = FALSE,
  ...
)
}
\arguments{
\item{x}{Object with class \code{tbl_summary} from the \link[gtsummary]{tbl_summary} function or
with class \code{tbl_svysummary} from the \link[gtsummary]{tbl_svysummary} function}

\item{statistic}{String indicating the statistic to report. Default is the
number of non-missing observation for each variable, \code{statistic = "{n}"}.
Other statistics available to report include:
\itemize{
\item \code{"{N}"} total number of observations,
\item \code{"{n}"} number of non-missing observations,
\item \code{"{n_miss}"} number of missing observations,
\item \code{"{p}"} percent non-missing data,
\item \code{"{p_miss}"} percent missing data
The argument uses \link[glue:glue]{glue::glue} syntax and multiple statistics may be reported,
e.g. \code{statistic = "{n} / {N} ({p}\%)"}
}}

\item{col_label}{String indicating the column label.  Default is \code{"**N**"}}

\item{footnote}{Logical argument indicating whether to print a footnote
clarifying the statistics presented. Default is \code{FALSE}}

\item{last}{Logical indicator to include N column last in table.
Default is \code{FALSE}, which will display N column first.}

\item{...}{Not used}
}
\description{
Add column with N
}
