% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstfun-package.R
\docType{package}
\name{bstfun-package}
\alias{bstfun}
\alias{bstfun-package}
\title{bstfun: Misc functions for the dept}
\description{
Collection of functions to include in biostatR that
    don't have a clear place to live the in more defined department
    packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/bstfun}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Margie Hannum (\href{https://orcid.org/0000-0002-2953-0449}{ORCID})
  \item Karissa Whiting (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})
}

}
\keyword{internal}
