% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstfun-package.R
\docType{package}
\name{bstfun-package}
\alias{bstfun}
\alias{bstfun-package}
\title{bstfun: Misc functions for the dept}
\description{
A miscellaneous collection of functions to used by members of the Biostatistics Department at MSKCC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/bstfun}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Jessica A. Lavery
  \item Karissa Whiting (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})
}

Other contributors:
\itemize{
  \item Margie Hannum (\href{https://orcid.org/0000-0002-2953-0449}{ORCID}) [contributor]
  \item Meier Hsu [contributor]
  \item Caroline Kostrzewa [contributor]
  \item Jasme Lee [contributor]
  \item Amy Tin (\href{https://orcid.org/0000-0002-8005-0694}{ORCID}) [contributor]
  \item Emily Vertosick [contributor]
  \item Gustavo Zapata Wainberg (\href{https://orcid.org/0000-0002-2524-3637}{ORCID}) [contributor]
  \item Christine Zhou [contributor]
}

}
\keyword{internal}
