% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biostat_project.R
\name{create_project}
\alias{create_project}
\alias{create_bst_project}
\title{Start a New Biostatistics project}
\usage{
create_bst_project(path, path_data = NULL, git = NA, ...)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created.}

\item{path_data}{A path. The directory where the secure data exist. Default is
\code{NULL}.  When supplied, a symbolic link to data folder will be created.}

\item{git}{Logical indicating whether to create Git repository.  Default is \code{TRUE}
When \code{NA}, user will be prompted whether to initialise Git repo.}

\item{...}{
  Arguments passed on to \code{\link[starter:create_project]{starter::create_project}}
  \describe{
    \item{\code{renv}}{Logical indicating whether to add renv to a project.
Default is \code{TRUE}. When \code{NA} user is asked interactively for preference.}
    \item{\code{symlink}}{Logical indicating whether to place a symbolic link
to the location in \verb{path_data=}. Default is to place the symbolic link
if the project is a git repository.}
    \item{\code{renv.settings}}{A list of renv settings passed to \code{renv::scaffold(settings=)}}
    \item{\code{overwrite}}{Logical indicating whether to overwrite existing files
if they exist.  Options are
\code{TRUE}, \code{FALSE}, and \code{NA} (aka ask interactively).  Default is \code{NA}}
    \item{\code{open}}{Logical indicating whether to open new project in fresh RStudio
session}
  }}
}
\description{
Creates a directory with the essential files for a new project.
The function can be used on existing project directories as well.

The folder name should be structured as \code{"<PI Last Name> <Short Description>"},
e.g. \code{"Sjoberg MRI detects Path Stage after Surgery"}. PI last name
is used in file naming.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# specifying project folder location (folder does not yet exist)
project_path <- fs::path(tempdir(), "My Project Folder")

# creating folder where secure data would be stored (typically will be a network drive)
secure_data_path <- fs::path(tempdir(), "secure_data")
dir.create(secure_data_path)

# creating new project folder
create_bst_project(project_path, path_data = secure_data_path)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[starter:create_project]{starter::create_project()}}
}
