% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_database_error.R
\name{fix_database_error}
\alias{fix_database_error}
\title{Function to make database fixes after import}
\usage{
fix_database_error(data, engine = readr::read_csv, ...)
}
\arguments{
\item{data}{data frame with errors}

\item{engine}{function to import file of database fixes}

\item{...}{arguments passed to the engine function to import the database fixes}
}
\value{
updated data frame
}
\description{
Function to make database fixes after import
}
\examples{
df_fixes <-
  tibble::tribble(
    ~id, ~variable, ~value,
    "id == 1", "age", "56",
    "id == 2", "trt", "Drug C"
  )
trial \%>\%
  dplyr::mutate(id = dplyr::row_number()) \%>\%
  fix_database_error(
    engine = I,
    x = df_fixes
  )
}
