% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followup_time.R
\name{followup_time}
\alias{followup_time}
\title{Report Follow-up Time Among Censored Obs}
\usage{
followup_time(
  Surv,
  data = NULL,
  pattern = "{median} (IQR {p25}, {p75})",
  style_fun = gtsummary::style_sigfig
)
}
\arguments{
\item{Surv}{An object of class \code{"Surv"} crated with \code{survival::Surv()}}

\item{data}{A data frame}

\item{pattern}{Statistics pattern to return. Default is
\code{"{median} (IQR {p25}, {p75})"}. User may select the following summary
statistics to report, \code{"{median}"}, \code{"{mean}"}, \code{"{p25}"}, \code{"{p75}"},
\code{"{sd}"}, \code{"{var}"}, \code{"{min}"}, and \code{"{max}"}.}

\item{style_fun}{Function used to style/format the summary statistics.
Default is \code{gtsummary::style_sigfig}. Argument accepts anonymous function
notation, e.g. \code{~gtsummary::style_sigfig(., digits = 3)}}
}
\value{
string of summary statistics
}
\description{
Function accepts \code{survival::Surv()} object, extracts the follow-up
times among the censored observations, and returns the requested
summary statistics. Use this function to report follow-up in-line in
R Markdown reports.
}
\examples{
library(survival)

followup_time(Surv(time, status), data = lung)

followup_time(
  Surv(time, status), data = lung,
  pattern = "{median} days",
  style_fun = ~gtsummary::style_sigfig(., digits = 4)
)
}
