% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalibration.R
\name{ggcalibration}
\alias{ggcalibration}
\title{Model Calibration Plot}
\usage{
ggcalibration(
  data,
  y,
  x,
  n.groups = 10,
  conf.level = 0.95,
  ci.method = c("exact", "ac", "asymptotic", "wilson", "prop.test", "bayes", "logit",
    "cloglog", "probit"),
  geom_smooth.args = list(method = "loess", se = FALSE, formula = y ~ x, color =
    "black"),
  geom_errorbar.args = list(width = 0),
  geom_point.args = list(),
  geom_function.args = list(colour = "gray", linetype = "dashed")
)
}
\arguments{
\item{data}{a data frame}

\item{y}{variable name of the outcome coded as 0/1}

\item{x}{variable name of the risk predictions}

\item{n.groups}{number of groups}

\item{conf.level}{level of confidence to be used in the confidence interval}

\item{ci.method}{method to use to construct the interval.
See \code{\link[binom:binom.confint]{binom::binom.confint()}} for details}

\item{geom_smooth.args}{named list of arguments that will be passed
to \code{ggplot2::geom_smooth()}. Default is
\code{list(method = "loess", se = FALSE, formula = y ~ x, color = "black")}}

\item{geom_errorbar.args}{named list of arguments that will be passed
to \code{ggplot2::geom_errorbar()}. Default is \code{list(width = 0)}}

\item{geom_point.args}{named list of arguments that will be passed
to \code{ggplot2::geom_point()}. Default is \code{list()}}

\item{geom_function.args}{named list of arguments that will be passed
to \code{ggplot2::geom_function()} and is the function that adds the 45 degree
guideline. Default is \code{list(colour = "gray", linetype = "dashed")}}
}
\value{
ggplot
}
\description{
Assess a model's calibration via a calibration plot.
}
\examples{
glm(response ~ age + marker + grade, trial, family = binomial) \%>\%
  broom::augment(type.predict = "response") \%>\%
  ggcalibration(y = response, x = .fitted, n.groups = 6) +
  ggplot2::xlim(0, 1) +
  ggplot2::labs(x = "Model Risk")
}
