% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_labels.R
\name{list_labels}
\alias{list_labels}
\title{Get variable labels and store in named list}
\usage{
list_labels(data)
}
\arguments{
\item{data}{Data frame}
}
\description{
Get variable labels and store in named list
}
\examples{
list_labels(trial)
}
\author{
Daniel D. Sjoberg
}
