% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_templates}
\alias{project_templates}
\title{Biostatistics project templates}
\format{
A named list containing the Biostatistics project template.
}
\usage{
project_templates
}
\description{
The \code{project_templates} object defines the contents of the Biostatistics project
templates used in \code{create_bst_project()} and \code{use_bst_file()}.
}
\examples{
if (FALSE) {
create_hot_project(
  path = file.path(tempdir(), "Sjoberg New Project"),
  template = bstfun::project_templates[["default"]]
)
}
}
\seealso{
\code{\link[=create_bst_project]{create_bst_project()}}

\code{\link[=use_bst_file]{use_bst_file()}}
}
\keyword{datasets}
