% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{style_pvalue}
\alias{style_percent}
\alias{style_sigfig}
\alias{style_ratio}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{all_of}
\alias{any_of}
\alias{everything}
\alias{last_col}
\alias{one_of}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gtsummary}{\code{\link[gtsummary]{style_percent}}, \code{\link[gtsummary]{style_pvalue}}, \code{\link[gtsummary]{style_ratio}}, \code{\link[gtsummary]{style_sigfig}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect]{any_of}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{last_col}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}}
}}

