% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{all_of}
\alias{any_of}
\alias{everything}
\alias{last_col}
\alias{one_of}
\alias{as_gt}
\alias{as_kable}
\alias{style_pvalue}
\alias{style_percent}
\alias{style_sigfig}
\alias{style_ratio}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr:reexports]{all_of}}, \code{\link[dplyr:reexports]{any_of}}, \code{\link[dplyr:reexports]{contains}}, \code{\link[dplyr:reexports]{ends_with}}, \code{\link[dplyr:reexports]{everything}}, \code{\link[dplyr:reexports]{last_col}}, \code{\link[dplyr:reexports]{matches}}, \code{\link[dplyr:reexports]{num_range}}, \code{\link[dplyr:reexports]{one_of}}, \code{\link[dplyr:reexports]{starts_with}}}

  \item{gtsummary}{\code{\link[gtsummary]{as_gt}}, \code{\link[gtsummary]{as_kable}}, \code{\link[gtsummary]{style_percent}}, \code{\link[gtsummary]{style_pvalue}}, \code{\link[gtsummary]{style_ratio}}, \code{\link[gtsummary]{style_sigfig}}}
}}

