% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_prior_pkgs.R
\name{reinstall_prior_pkgs}
\alias{reinstall_prior_pkgs}
\title{Re-install pkgs from a prior installation of R}
\usage{
reinstall_prior_pkgs(path = NULL)
}
\arguments{
\item{path}{Path to the package library from the previous installation of R.
Default is \code{NULL}. When \code{NULL}, the location of the library is inferred. If
the location cannot be inferred, user must provide the path.}
}
\value{
\code{NULL}
}
\description{
When install a new version of R, your package library is empty and
packages you use need to be re-installed. This function will help you
install all the packages you had previously installed in the new R package
library.
\itemize{
\item \emph{Function installs the current release of a pkg from CRAN or a secondary repository you may have set.} Packages installed from other locations, such as GitHub, will need to be re-installed manually.
\item The pkgs are installed using \code{renv::install()} adding each package to your renv cache
\item Function should be run in a fresh R Session outside of an RStudio project
\item Any packages already installed will be skipped
}
}
