% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_tbl_compact.R
\name{style_tbl_compact}
\alias{style_tbl_compact}
\title{Compact Table Styling}
\usage{
style_tbl_compact(data)
}
\arguments{
\item{data}{a {gt}, {flextable}, {huxtable}, or \code{knitr::kable()} table.}
}
\description{
Apply the same compact styling available from
\code{gtsummary::theme_gtsummary_compact()} to any
{gt}, {flextable}, {huxtable}, or \code{knitr::kable()} table.
\code{knitr::kable()} stlying uses the {kableExtra} package
}
\examples{
style_tbl_compact_ex1 <-
  head(trial) \%>\%
  gt::gt() \%>\%
  style_tbl_compact()
}
