% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_likert.R
\name{tbl_likert}
\alias{tbl_likert}
\title{Likert Summary}
\usage{
tbl_likert(
  data,
  label = NULL,
  statistic = NULL,
  digits = NULL,
  include = everything(),
  sort = c("default", "ascending", "descending")
)
}
\arguments{
\item{data}{A data frame}

\item{label}{List of formulas specifying variables labels,
e.g. \code{list(age ~ "Age", stage ~ "Path T Stage")}.  If a
variable's label is not specified here, the label attribute
(\code{attr(data$age, "label")}) is used.  If
attribute label is \code{NULL}, the variable name will be used.}

\item{statistic}{Formula or list of formulas specifying types of categorical
statistics to display, see \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}} help page,
section \emph{statistic argument}}

\item{digits}{Formula or list of formulas indicating how to display the
computed statistics, see \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}} help page}

\item{include}{variables to include in the summary table. Default is \code{everything()}}

\item{sort}{Sort table based on mean scores? Must be one of
\code{c("default", "ascending", "descending")}}
}
\description{
\lifecycle{experimental}
}
\examples{
df <-
  tibble::tibble(
    f1 =
      sample.int(100, n = 3, replace = TRUE) \%>\%
      factor(levels = 1:3, labels = c("bad", "meh", "good")),
    f2 =
      sample.int(100, n = 3, replace = TRUE) \%>\%
      factor(levels = 1:3, labels = c("bad", "meh", "good")),
  )

tbl_likert(df) \%>\%
  add_n() \%>\%
  gtsummary::as_kable()
}
