% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtsummary_msk.R
\name{theme_gtsummary_msk}
\alias{theme_gtsummary_msk}
\title{Set custom gtsummary themes}
\usage{
theme_gtsummary_msk(
  name = c("hot", "karissa", "ally", "mauguen", "esther", "curry", "lavery", "meier",
    "leej", "whitingk", "mauguena", "drille", "currym1", "laveryj", "hsum1", "leej22"),
  font_size = NULL
)
}
\arguments{
\item{name}{string indicating the custom theme to set.}

\item{font_size}{font size passed to compact them}
}
\description{
This is a place for any member of the MSK community to add a personal gtsummary theme.
Reach out if you're interested in adding yours!
}
\details{
Visit the \href{http://www.danieldsjoberg.com/gtsummary/articles/themes.html#writing-themes-1}{gtsummary themes vignette} for a full list of preferences that can be set.
}
\examples{
theme_gtsummary_msk("hot")
}
\seealso{
Other gtsummary-related functions: 
\code{\link{add_inline_forest_plot}()},
\code{\link{add_sparkline}()},
\code{\link{as_ggplot}()},
\code{\link{bold_italicize_group_labels}()},
\code{\link{logistic_reg_adj_diff}()},
\code{\link{style_tbl_compact}()},
\code{\link{tbl_likert}()}
}
\concept{gtsummary-related functions}
