% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bst_rstudio_prefs.R
\name{use_bst_rstudio_prefs}
\alias{use_bst_rstudio_prefs}
\title{Use Biostatistics RStudio Preferences}
\usage{
use_bst_rstudio_prefs()
}
\description{
The function wraps \code{rstudio.prefs::use_rstudio_prefs()} and sets the following
preferences in RStudio.\tabular{ll}{
   \strong{Preference} \tab \strong{Value} \cr
   always_save_history \tab FALSE \cr
   graphics_backend \tab ragg \cr
   load_workspace \tab FALSE \cr
   margin_column \tab 80 \cr
   rainbow_parentheses \tab TRUE \cr
   restore_last_project \tab FALSE \cr
   rmd_chunk_output_inline \tab FALSE \cr
   show_hidden_files \tab TRUE \cr
   show_invisibles \tab TRUE \cr
   show_last_dot_value \tab TRUE \cr
   show_line_numbers \tab TRUE \cr
   show_margin \tab TRUE \cr
   save_workspace \tab never \cr
}
}
\examples{
\dontrun{\donttest{
use_bst_rstudio_prefs()
}}
}
