% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bst_rstudio_prefs.R
\name{use_bst_rstudio_prefs}
\alias{use_bst_rstudio_prefs}
\title{Use Biostatistics RStudio Preferences}
\usage{
use_bst_rstudio_prefs()
}
\description{
The function wraps \code{rstudio.prefs::use_rstudio_prefs()} and sets the following
preferences in RStudio.<table>
 <thead>
  <tr>
   <th style="text-align:left;"> **Preference** </th>
   <th style="text-align:left;"> **Value** </th>
  </tr>
 </thead>
<tbody>
  <tr>
   <td style="text-align:left;"> always_save_history </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> graphics_backend </td>
   <td style="text-align:left;"> ragg </td>
  </tr>
  <tr>
   <td style="text-align:left;"> load_workspace </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> margin_column </td>
   <td style="text-align:left;"> 80 </td>
  </tr>
  <tr>
   <td style="text-align:left;"> rainbow_parentheses </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> restore_last_project </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> rmd_chunk_output_inline </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> show_hidden_files </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> show_invisibles </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> show_last_dot_value </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> show_line_numbers </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> show_margin </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
  <tr>
   <td style="text-align:left;"> save_workspace </td>
   <td style="text-align:left;"> never </td>
  </tr>
</tbody>
</table>
}
\examples{
\dontrun{\donttest{
use_bst_rstudio_prefs()
}}
}
